refactor: simplify extension
