+++ apps/extension/src/background.ts
