+++ apps/extension/src/components/Avatar.tsx
