+++ apps/extension/src/components/AvatarList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { StyledComponentPropsWithRef } from 'styled-components';
 import { variant } from 'styled-system';
 import { Avatar } from './Avatar';
-import { IUser } from '../../types';
+import { IUser } from '../types';
 
 type Variants = 'row' | 'col';
