+++ apps/extension/src/components/HomeHighlights.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { IHighlight } from 'colighter';
-import { useCollabHighlights } from '../../utils';
+import { useCollabHighlights } from '../utils';
 
 const Text = styled.p<{ isFirst: boolean }>`
   text-align: left;
