+++ apps/extension/src/types.ts
@@ -1,13 +1,7 @@
-export interface ColorDescription {
-  name: string;
-  val: string;
-}
-
 export enum MessageAction {
   TOGGLE_HIGHLIGHTS = 'TOGGLE_HIGHLIGHTS',
   REMOVE_HIGHLIGHTS = 'REMOVE_HIGHLIGHTS',
   RENDER_HIGHLIGHTS = 'RENDER_HIGHLIGHTS',
-  SELECT_COLOR = 'SELECT_COLOR',
   LOAD_COLLAB = 'LOAD_COLLAB',
   GET_COLLAB_HIGHLIGHTS = 'GET_COLLAB_HIGHLIGHTS',
   POST_COLLAB_HIGHLIGHTS = 'POST_COLLAB_HIGHLIGHTS',
@@ -43,13 +37,3 @@ export interface IUser {
   userName: string;
   imageUrl: string;
 }
-
-export const HIGHLIGHT_COLOR_OPTIONS: ColorDescription[] = [
-  { name: 'red', val: 'FAA99D' },
-  { name: 'yellow', val: 'FDDF7E' },
-  { name: 'green', val: 'CCE29C' },
-  { name: 'blue', val: '67EBFA' },
-  { name: 'purple', val: 'CE97FB' },
-];
-
-export const DEFAULT_HIGHLIGHT_COLOR = HIGHLIGHT_COLOR_OPTIONS[0];
