--- apps/extension/src/view/theme/index.ts
@@ -1,1 +1,0 @@
-export * from './Theme';
